update: legacy types
