+++ src/types/api/accompanying.ts
@@ -1,1 +1,10 @@
-export type ApiOpportunityAccompanyingDetails = Record<string, unknown>;
+import { TranslatedIntoType } from "./opportunity";
+
+export interface ApiOpportunityAccompanyingDetails {
+  appointmentAddress?: string;
+  appointmentDate?: Date;
+  appointmentTime?: Date;
+  refugeeNumber?: string;
+  refugeeName?: string;
+  languageToTranslate?: TranslatedIntoType;
+}
